Refactored back to working state, Second ajax function displays  campgrounds to page
