+++ client/components/initialView.js
@@ -1,0 +1,13 @@
+
+//initial app view
+
+// var React = require('react');
+
+// var HelloMessage = React.createClass({
+//   render: function() {
+//     return <div>Hello {this.props.name}</div>;
+//   }
+// });
+
+// module.exports = HelloMessage;
+
